+++ src/chip/imxrt10xx.rs
@@ -14,6 +14,9 @@ cfg_if::cfg_if! {
     if #[cfg(chip = "imxrt1010")] {
         #[path = "imxrt10xx/imxrt1010.rs"]
         pub(crate) mod config;
+    } else if #[cfg(chip = "imxrt1020")] {
+        #[path = "imxrt10xx/imxrt1020.rs"]
+        pub(crate) mod config;
     } else if #[cfg(any(chip = "imxrt1060", chip = "imxrt1064"))] {
         #[path = "imxrt10xx/imxrt1060.rs"]
         pub(crate) mod config;