+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+Add support for i.MX RT 1020 processors. Enable support with the `"imxrt1020"`
+imxrt-hal feature, and separately select your imxrt-ral feature.
+
 ## [0.5.0] 2022-01-05
 
 `imxrt-hal` provides common peripherals for all chips supported by `imxrt-ral`,